Refactor image classifications into separate library class. #40
